refactor: Move json related functions to monkey_patch_json
